Recover testbed via console.  (#10806)
What is the motivation for this PR?
There are lots of unhealthy testbeds in our pool, which needed to be recovered. But some of them lost their management ip, and could not be connected via ssh. So we need another method to access these devices and recover them. In this PR, we connect to devices via console, and try to recover it via console.

In this first phase, the script can be used to recover unhealthy testbeds using ONIE and ABOOT. And other types of testbeds will be supported in next phase.

How did you do it?
In this PR, we will check if we can connect to the device via ssh and if sonic-installer is usable on the device. If not, we will connect to the device via console and do power cycle.

We should execute command like

python3 ../.azure-pipelines/recover_testbed/recover_testbed.py -i <inventory> <inventory> -t <testbed name> --tbfile <testbed file> --log-level info --image <image url> --hwsku <hwsku>
under folder sonic-mgmt/ansible

How did you verify/test it?
I tested on several testbeds and these testbeds can be recovered via my script.