Fit celestica (#11809)
Approach
What is the motivation for this PR?
This PR addresses the issue of celestica testbeds getting stuck during ONIE booting. We found that sending an enter to the terminal can prevent this problem. Therefore, we added a step to send an enter while recovering celestica testbeds in ONIE mode.

How did you do it?
Add a step to send an enter while recovering celestica testbeds in ONIE mode.