[devutils] Fix issue of parsing output_watts of PDU (#11831)
What is the motivation for this PR?
The devutils tool may raise exception like below if the output_watts of PDU is empty string '':

ValueError("invalid literal for int() with base 10: ''")

How did you do it?
This change improved the code of parsing output_watts. Another issue is that a temporary variable named status conflicts with another variable. This change also fixed that.

How did you verify/test it?
Verified by trying to get pdu status using the devutils tool.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>