Add topology mark in the module already in PR test (#12390)
What is the motivation for this PR?
In our pytest, it's mandatory to have proper markers for different test topologies to ensure that tests are run in the correct testbeds. It seems that some modules, despite having passed the PR tests, lack the necessary topology markers. In this PR, we add the missing markers, which will enhance the clarity and efficiency of the testing process and ensure that each module is tested on the appropriate testbeds.