[override] Aadd ASIC_SENSORS table to non-user config table (#12545)
## Description of PR

Summary: Add ASIC_SENSORS to non-user-config table

## Approach
#### What is the motivation for this PR?
The ASIC_SENSORS table is not generated from minigraph but the derivative config after reload. This will cause random error if we take the screenshot before and after its occurence thus result in error of 8111 platform. 
#### How did you do it?
Ignore this table in comparison.
#### How did you verify/test it?
E2E