Get the status of github issue without github authentication.  (#12790)
What is the motivation for this PR?
In conditional mark, we may need to query the status of github issue. We use personal github authentication token to do this before. But per to the official document, "This endpoint can be used without authentication or the aforementioned permissions if only public resources are requested." So due to the security, we remove the github authentication here.

How did you do it?
Remove the github authentication when querying the status of github issue.