Change branch (#13355)
Description of PR
This is a bugfix of PR #13321. Previously, we set the wrong branch. The pipeline will run on public repo, so we will use master branch instead of internal.

Approach
What is the motivation for this PR?
This is a bugfix of PR #13321. Previously, we set the wrong branch. The pipeline will run on public repo, so we will use master branch instead of internal.

co-authorized by: jianquanye@microsoft.com