[tests/copp]: Fix copp tests (#12836)
This PR does the following:

Enhance copp test cases so that they pass consistently on Broadcom platforms.
For Broadcom platforms, when BGP and IP2ME traps are removed, BGP packets hit the default trap group, so a new
DefaultPolicy is introduced in check_constraints specifically for Broadcom ASIC types. The test logic for non-Broadcom
platforms is left unchanged. A necessary side change of this is that the CIR for default trap needs to be changed to a
value other than 600 only for the test. This is reverted back at the end of the test.
Remove xfail for copp tests on broadcom platforms
Skip updating default sonic config for queue4_group3 in scripts/update_copp_config.py as the new rate-limit for this copp
group is lower than the default value of 600 PPS used by copp tests. This value is 300PPS for Mgmt devices and 100 PPS for
other devices. The constraint checks for protocols that map to queue4_group3 are also changed accordingly in
ansible/roles/test/files/ptftests/py3/copp_tests.py
Summary:
Fixes # 27262099

What is the motivation for this PR?
When BGP and IP2ME traps are removed, BGP packets hit the default trap group on broadcom platforms and trap to CPU queue0. This change modifies copp test to verify this behavior.

How did you do it?
By changing the rate limit used in the test for default trap group, it is easy to verify that packets are hitting that trap group. If all trap groups use the same rate limit of 600 pps as is done currently, it is rather difficult to identify which trap group got hit.

How did you verify/test it?
By running tests locally and using azure pipeline