Add script `test_snmp_psu.py` into T0 PR checker (#13810)
What is the motivation for this PR?
In this PR, we optimize the script test_snmp_psu.py to support running on kvm testbed.
Command psuutil numpsus will get the expected return code 2 on kvm testbed because of no chassis. So, on kvm testbed, we will return in advance.
There is no snmp psu infomation of kvm testbed. So we return in advance to skip the following checks.

How did you do it?
Return in advance on kvm testbed to skip following checks.