Optimize test_platform_serial_no to handle module sonic_platform variations (#13835)
What is the motivation for this PR?
In PR #13424, we have optimized the code to fit for kvm testbed. But I notice that, in some kvm images, it includes the module sonic_platform, but in other images, they doesn't include. This will generate different return code while executing some commands. In this PR, we optimize the code to cover both situations.

How did you do it?
Optimize the code to cover different situations.