[Snappi] Fix for crash in enable_packet_aging_after_test called during test cleanup (#13845)
Description of PR
Function enable_packet_aging_after_test is defined in conf_test.py to be called as test-cleanup after test execution. This function calls enable_packet_aging defined in common/snappi_tests/common_helper.py. However, it does not pass the asic_value when calling this function. There is no check in enable_packet_aging function to handle this and it crashes.

Summary:
Fixes # (issue)
#13654

Approach
What is the motivation for this PR?
To fix the enable_packet_aging during the test-cleanup. There is no asic_value defined when this function is called during test-cleanup and function does not have any check to handle this.

How did you do it?
Added check to see if the asic_value is present or not. If the asic_value is present, then default code is executed. If the asic_value is not present, then function will enable packet aging for all ASICs present on the DUT.

Note - By default, packet aging is enabled on the DUT and hence enabling it on all ASICs during the cleanup should be ok.

How did you verify/test it?
Normal OC runs, the test clean up works fine.

co-authorized by: jianquanye@microsoft.com