Prevent duplication of topology subclasses in PR statistic (#14474)
What is the motivation for this PR?
In PR #14437, we consider the topologies m0, mx and mc0 as subclassses of t0. Since we have scripts marked for both t0 and its subclasses, we add a condition in this PR to avoid duplication in our statistics.

How did you do it?
Add a condition in this PR to avoid duplication in our statistics.