[dhcp_server_test] fix index error because of no ip left in list (#14477)
When selecting two ip for a range, we first pick one ip as left boundary and pick the second ip as right boundary.
to make sure the range is valid, we pick the second ip in range from the ip next to left boundary to the end of the list.
However if the last ip was chosed randomly, the rest ip will be empty, and raise index error. we need to add boundary when selecting ip randomly.