[platform_tests/counterpoll]: Don't run counterpoll_watermark test on supervisor (#14724)
Select only frontend hwskus for counterpoll_queue_watermark_pg_drop test
Assert if waiting for COUNTER_DB to be populated fails
BugLink: #14721

Summary:
Fixes #14721

Approach
What is the motivation for this PR?
platform test test_counterpoll_watermark_pg_drop runs and fails on the sup cards, where it should only be running on linecards. Test case fixture needs to be modified to only select frontend cards. Additionally, the test should fail if it fails to wait on the COUNTER_DB on becoming populated

How did you do it?
How did you verify/test it?
Run on testbed - successfully skips supervisor, and asserts properly if the counter_db is not populated

Signed-off-by: Liam Kearney <liamkearney@microsoft.com>