Rename snmp/test_snmp_link_local.py in pr test (#15175)
What is the motivation for this PR?
In PR #9268, a test script named test_snmp_link_local.py was introduced, but it was incorrectly listed as snmp/test_snmp_link_local_ip.py in .azure-pipelines/pr_test_scripts.yaml. This PR corrects the entry in .azure-pipelines/pr_test_scripts.yaml to resolve the mismatch.

How did you do it?
This PR corrects the entry in .azure-pipelines/pr_test_scripts.yaml to resolve the mismatch.

How did you verify/test it?