sonic-mgmt: fix info in testbed_vm_info.py err msg (#15249)
In ansible/library/testbed_vm_info.py, there is a
format string passed to a method without the
actual formatting applied, i.e. the '{}'s are
printed out rather than the desired information
they're supposed to represent. This change
passes the error message to the method with the
correct formatting applied.