[advanced-reboot] Test VLAN gateway address resolvability during upgrade (#15835)
What is the motivation for this PR?
During an upgrade, test to see if the VLAN gateway can be resolved with an ARP request. When CPA is enabled, we expect the VLAN gateway to be always resolvable (even when control plane of the device is down). This is done through the use of a Ferret server running on the PTF container.

How did you do it?
Add a basic check for the functionality of that Ferret server by crafting and sending ARP requests, and report the number of ARP responses received. As of right now, this doesn't affect the final result of the warm/fast upgrade test; that will be done later.

Also mark the advanced-reboot tests as functional on KVM. Without this, the warm upgrade test will not run on PR checkers.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>