Relocate Impacted Area-Based PR Testing HLD (#16484)
What is the motivation for this PR?
In PR #14761, we introduced the HLD for Impacted Area-Based PR testing. However, the document was placed in an outdated path. In this PR, we relocate it to the correct path.

How did you do it?
In this PR, we relocate the HLD to the correct path.