[Impacted Area Based PR testing] Roll out rest of the PR checkers (#16565)
What is the motivation for this PR?
In #15666, we introduced a new approach to PR testing called Impacted Area-Based PR Testing. This model will be rolled out in phases. This PR implement the left PR checkers.

How did you do it?
Roll out the left PR checkers.

How did you verify/test it?
Test by pipeline itself, to see if PR checkers are running as expected.