Skip test_srv6_static_config.py in PR testing (#16629)
What is the motivation for this PR?
Script srv6/test_srv6_static_config.py was added in PR #16516, and it fails to run on kvm testbed. We temporarily skip this test in PR testing using conditional mark.

How did you do it?
We temporarily skip this test in PR testing using conditional mark.

How did you verify/test it?
Test by pipeline itself to see if this script can be skipped on kvm testbed.