Add support for testing chrony (#15008)
* Add support for bringing up a DUT with chrony instead of ntpd

When running deploy-mg on a device that has chrony instead of ntpd, the
steps to synchronize the time are different. Check to see if chrony is
available, and if so, use different steps to synchronize the time.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Add support for testing chrony in the NTP tests

In ntp/test_ntp.py, add support for testing the chrony daemon instead of
ntpd. Additionally, for the long jump disabled test case, make sure that
the time isn't synchronized when using chrony.

---------

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>