[Impacted Area Based PR testing] Enhancing regex accuracy in impacted area detection. (#16917)
What is the motivation for this PR?
In the Get Impacted Area stage, we use regular expressions to determine whether the changes in a PR are related to a specific feature. In this PR, we refine the regular expression to reduce false matches and improve accuracy.

How did you do it?
In this PR, we refine the regular expression to reduce false matches and improve accuracy.

How did you verify/test it?
For string like ansible/roles/files/ptftests, we can't match the pattern. And for string start with tests/, we can match this pattern.