Fix baseline tests (#17089)
What is the motivation for this PR?
This PR fixes a bug in the recent refactor to baseline tests. Fix passes the correct variable to run the test scripts.

How did you do it?
Pass the correct variable to run the test scripts.

How did you verify/test it?
Run baseline test pipeline.