Refine the logic of matching conditional mark. (#16930)
What is the motivation for this PR?
In PR #14395, we optimized the matching rule for conditional marks by considering all potential matches. If a mark is unique among these matches, it is applied to the test case. Otherwise, the longest matching entry is used.

In this PR, we further refine this logic by ensuring that the longest match where the condition is true is used. This means that if the conditions of the longest matching entry are false, we will consider the next longest matching entry.

How did you do it?
How did you verify/test it?