[test_baud_rate_boot_connect] skip memory utilization (#17389)
What is the motivation for this PR?
The case would do the reboot and send "ctrl + C" to let the device run into ABoot.
Then do the reboot again in the teardown to recover the device.
However, the memory check teardown would be called before the case module teardown.

How did you do it?
Skip the memory check for this case test_baud_rate_boot_connect

How did you verify/test it?
Local run the case, skipped the memory check for this case