Add the correct queue to select for test_snappi (#18018)
Description of PR
Summary:
tests/snappi_tests/test_snappi.py is failing when we use the non-default queue mapping in variables.py. This is because it uses the priority as the queue id.
In this PR we change the code to use the correct queue_id from the variables.py.

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 New Test case
 Skipped for non-supported platforms
 Test case improvement
Back port request
 202012
 202205
 202305
 202311
 202405
 202411
Approach
What is the motivation for this PR?
The failure of test_snappi.py for non default queue mapping.

How did you do it?
Pls see description.

How did you verify/test it?
Ran it in T2 testbed:

co-authorized by: jianquanye@microsoft.com