Handling the scenario when number of snappi ports used and number of l3 interfaces or vlan members on DUT are not same (#17966)
Description of PR
When the number if snappi ports used in the test and number of l3 interfaces or number of Vlan members configured on DUT are not same it was failing to generate the snappi configuration and was failing in

https://wwwin-github.cisco.com/whitebox/sonic-test/blob/343dc8e4dfb9d6137e6ff09cc9a9a1b516e76daf/sonic-mgmt/tests/common/snappi_tests/snappi_fixtures.py#L152

OR

https://wwwin-github.cisco.com/whitebox/sonic-test/blob/343dc8e4dfb9d6137e6ff09cc9a9a1b516e76daf/sonic-mgmt/tests/common/snappi_tests/snappi_fixtures.py#L233

Summary:
Fixes the error seen and generates the Snappi configuration accordingly.

Type of change
 Bug fix
[X ] Testbed and Framework(new/improvement)
 New Test case
 Skipped for non-supported platforms
 Test case improvement
Back port request
 202012
 202205
 202305
 202311
 202405
[X ] 202411
Approach
What is the motivation for this PR?
Tests were failing

How did you do it?
Added the fix to handle the scenario

How did you verify/test it?
Manually ran the test with changes and verified it passes

Any platform specific information?
No

Supported testbed topology if it's a new test case?
No

co-authorized by: jianquanye@microsoft.com