[DualToR] follow-up to #18766: add missing iptables rule (#18920)
What is the motivation for this PR?
This is a follow-up to #18766, which enforced same loopback ip and drop bgp pkts on loopback1 for dualtor.
However, I noticed that one commit related to appending the iptables/ip6tables rule was not included in the previous PR. This PR adds the missing iptables rule to ensure that is correctly applied.

How did you do it?
Append the missing commit.

How did you verify/test it?
It was verified in #18766