Fix gcu/test_cacl for multi-asic t1 (#18361)
Summary: this is a part of the initiative that tries to fix existing T1 test cases that fail on multi-asic platforms.
This PR fixes generic_config_updater/test_cacl.py: it failed on multi-asic T1 because previously the author of the test case did not format the GCU patches with namespace properly.

* remove changes in test_k8s_config