[dualtor] Skip toggle failure in `test_container_checker` (#20183)
What is the motivation for this PR?
test_container_checker has toggle failure introduced by setup_dualtor_mux_ports.
the toggle failure is due to the test will stop mux container, so setup_dualtor_mux_ports will fail to toggle in the following case as the mux is not running.

Signed-off-by: Longxiang Lyu <lolv@microsoft.com>