[test report] Improve robustness of junit xml parser (#20461)
What is the motivation for this PR?
The junit xml parser is not able to handle some special cases like a test script with no case executed or skipped.

How did you do it?
This change improved the junit xml parser for the corner cases.

When test suite has no children, testsuit_element is not None, but if not testsuit_element evaluates to True. Update the code to use explicit if testsuit_element is None.
When no case is executed or skipped, the tag may still have a bare child like <testcase time="17.190" />. It will fail at getting mandatory attributes. This change improved the logic for this special case to return None.
Improved debuggability by including junit xml file name in the roots list. Then we can add print to check exception is raised for which junit xml file.
How did you verify/test it?
Test run junit_xml_parser.py and report_uploader.py.